refactor : breadcrumb 눌렀을때 이벤트 처리 코드 정리
